uRo3YA/requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
cryptography==38.0.4
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.13
django-allauth==0.51.0
django-appconf==1.0.5
django-bootstrap5==22.1
django-cors-headers==3.13.0
django-extensions==3.2.1
django-imagekit==4.1.0
django-shortcuts==1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.1.1
idna==3.4
ipython==8.5.0
jedi==0.18.1
Markdown==3.4.1
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pickleshare==0.7.5
pilkit==2.0
Pillow==9.2.0
prompt-toolkit==3.0.31
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
traitlets==5.4.0
urllib3==1.26.13
wcwidth==0.2.5
